--- a/beater/pub.go
+++ b/beater/pub.go
@@ -26,8 +26,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/elastic/apm-agent-go"
-	"github.com/elastic/apm-server/config"
-	"github.com/elastic/apm-server/model"
+	"github.com/elastic/apm-server/transform"
 	"github.com/elastic/beats/libbeat/beat"
 )
 
@@ -47,9 +46,9 @@ type publisher struct {
 }
 
 type pendingReq struct {
-	payload model.Payload
-	config  config.Config
-	trace   bool
+	transformables []transform.Eventable
+	tcontext       transform.Context
+	trace          bool
 }
 
 var (
@@ -143,8 +142,12 @@ func (p *publisher) processPendingReq(req pendingReq) {
 		defer tx.End()
 	}
 
+	var events []beat.Event
 	span := tx.StartSpan("Transform", "Publisher", nil)
-	events := req.payload.Transform(req.config)
+	for _, transformable := range req.transformables {
+		events = append(events, transformable.Events(&req.tcontext)...)
+	}
+
 	span.End()
 
 	span = tx.StartSpan("PublishAll", "Publisher", nil)