--- a/beater/pub.go
+++ b/beater/pub.go
@@ -46,9 +46,9 @@ type publisher struct {
 }
 
 type pendingReq struct {
-	transformables []transform.Eventable
-	tcontext       transform.Context
-	trace          bool
+	eventables []transform.Eventable
+	tcontext   transform.Context
+	trace      bool
 }
 
 var (
@@ -144,8 +144,8 @@ func (p *publisher) processPendingReq(req pendingReq) {
 
 	var events []beat.Event
 	span := tx.StartSpan("Transform", "Publisher", nil)
-	for _, transformable := range req.transformables {
-		events = append(events, transformable.Events(&req.tcontext)...)
+	for _, eventable := range req.eventables {
+		events = append(events, eventable.Events(&req.tcontext)...)
 	}
 
 	span.End()