… config/config.go
… model/context.go
… model/context_test.go
… model/error/event_test.go
… model/error/payload_test.go
… model/metadata/metadata.go
… model/metadata/metadata_test.go
… model/metadata/process_test.go
… model/metadata/service_test.go
… model/metadata/system_test.go
… model/metadata/user.go
… model/metadata/user_test.go
… model/metric/event.go
… model/metric/event_test.go
… model/metric/payload.go
… model/metric/payload_test.go
… model/sourcemap/payload.go
… model/sourcemap/payload_test.go
… model/span/span.go
… model/span/span_context_test.go
… model/span/span_test.go
… model/stacktrace.go
… model/stacktrace_frame.go
… model/stacktrace_frame_test.go
… model/stacktrace_test.go
… model/system.go
… model/transaction/event.go
… model/transaction/event_test.go
… model/transaction/payload.go
… model/transaction/payload_test.go