* `config.Config` is now `transform.Config`
* `decoder.Decoder` is now `decoder.ReqDecoder` because it decodes http.Request objects which is different from the utility.ManualDecoder which operates on map[string]interface{}.
* All transform methods now take a `*transform.Context`. `transform.Context` has a `transform.Config` and a `Metadata`.
* `Metadata` has Service/Process/System/User
* `PayloadProcessor` is now `EventsProcessor`
* Individual `DecodePayload` methods are now abstracted into `EventsProcessor`
* pendingReq takes `[]transform.Transformable` instead of model.Payload.
* the Payload structs are gone. We have `Metadata` and `Transformable` instead. For sourcemaps I have just renamed it from "Payload" to "Sourcemap".
* Metadata has `Merge` and `MergeMininal` to let it merge into context for transaction and span contexts.
* `model.Context` and `span.SpanContext` are gone
* Metrics `apm-server.processor.{transaction,span,error,sourcemap}.transformation` now counts individual event transformation instead of one for each payload transformation.
* Spans have gained a timestamp and transaction ID. This was necessary in order to do the transformation individually. It's set to the values of the parent transaction at the moment. 

Removing the payloads here should enable us to do intake v2 in the next PR.