@roncohen 2018-07-24
Here are the benchmarks.
I saw rather big changes in the "wall clock" benchmarks (the ones that measure `ns` and not allocations) between the two branches. So I ran the benchmarks twice on `master` and compared and also saw very large difference, so it seems we can't trust those too much. However the allocations benchmarks are very interesting. 

Most tests have less allocations, except for the following. I'll investigate: 
```
benchmark                                                                             old bytes     new bytes     delta
BenchmarkBackendProcessor/TestProcessErrorFullTransform-4                             16913         25523         +50.91%
BenchmarkBackendProcessor/TestProcessErrorFullProcessRequest-4                        22837         31362         +37.33%
BenchmarkBackendProcessor/TestProcessErrorNullValuesTransform-4                       7840          12176         +55.31%
BenchmarkBackendProcessor/TestProcessErrorNullValuesProcessRequest-4                  9385          13626         +45.19%
BenchmarkBackendProcessor/TestProcessTransactionFullTransform-4                       24505         37905         +54.68%
BenchmarkBackendProcessor/TestProcessTransactionFullProcessRequest-4                  33465         47002         +40.45%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesTransform-4                 14640         21160         +44.54%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesProcessRequest-4            18612         25171         +35.24%

```

 
<details>


```
benchmark                                                                             old ns/op     new ns/op     delta
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceValidate-4                    8779          9925          +13.05%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceDecode-4                      4151          4159          +0.19%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceTransform-4                   8386          6480          -22.73%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceProcessRequest-4              21758         17099         -21.41%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessValidate-4                    11468         11928         +4.01%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessDecode-4                      4939          5219          +5.67%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessTransform-4                   8570          7373          -13.97%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessProcessRequest-4              20822         22390         +7.53%
BenchmarkBackendProcessor/TestProcessErrorFullValidate-4                              50229         54752         +9.00%
BenchmarkBackendProcessor/TestProcessErrorFullDecode-4                                15601         18001         +15.38%
BenchmarkBackendProcessor/TestProcessErrorFullTransform-4                             39793         49969         +25.57%
BenchmarkBackendProcessor/TestProcessErrorFullProcessRequest-4                        104403        113710        +8.91%
BenchmarkBackendProcessor/TestProcessErrorNullValuesValidate-4                        26709         30525         +14.29%
BenchmarkBackendProcessor/TestProcessErrorNullValuesDecode-4                          8531          9152          +7.28%
BenchmarkBackendProcessor/TestProcessErrorNullValuesTransform-4                       20707         26279         +26.91%
BenchmarkBackendProcessor/TestProcessErrorNullValuesProcessRequest-4                  50394         58619         +16.32%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPValidate-4                       9604          10927         +13.78%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPDecode-4                         4777          4596          -3.79%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPTransform-4                      8812          7893          -10.43%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPProcessRequest-4                 19591         19839         +1.27%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionValidate-4           4189          4589          +9.55%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionDecode-4             2981          3190          +7.01%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionTransform-4          7070          6055          -14.36%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionProcessRequest-4     12812         13425         +4.78%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogValidate-4                 6290          7209          +14.61%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogDecode-4                   2998          3037          +1.30%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogTransform-4                7141          6397          -10.42%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogProcessRequest-4           14456         14003         -3.13%
BenchmarkRumProcessor/TestProcessErrorRumValidate-4                                   25727         26550         +3.20%
BenchmarkRumProcessor/TestProcessErrorRumDecode-4                                     9314          10062         +8.03%
BenchmarkRumProcessor/TestProcessErrorRumTransform-4                                  726074        816179        +12.41%
BenchmarkRumProcessor/TestProcessErrorRumProcessRequest-4                             760458        795659        +4.63%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapValidate-4                             23783         23911         +0.54%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapDecode-4                               8067          8696          +7.80%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapTransform-4                            577903        615804        +6.56%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapProcessRequest-4                       643431        642055        -0.21%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapValidate-4                       28751         31146         +8.33%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapDecode-4                         9299          9832          +5.73%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapTransform-4                      74851         83283         +11.27%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapProcessRequest-4                 111740        128217        +14.75%
BenchmarkRumProcessor/TestProcessErrorAugmentedUserAgentAndIPValidate-4               9171          9970          +8.71%
BenchmarkRumProcessor/TestProcessErrorAugmentedUserAgentAndIPDecode-4                 4601          4805          +4.43%
BenchmarkRumProcessor/TestProcessErrorAugmentedUserAgentAndIPTransform-4              8615          7134          -17.19%
BenchmarkRumProcessor/TestProcessErrorAugmentedUserAgentAndIPProcessRequest-4         19897         19093         -4.04%
BenchmarkProcessor/TestProcessMetricValidate-4                                        25763         29386         +14.06%
BenchmarkProcessor/TestProcessMetricDecode-4                                          6558          7322          +11.65%
BenchmarkProcessor/TestProcessMetricTransform-4                                       10464         11929         +14.00%
BenchmarkProcessor/TestProcessMetricProcessRequest-4                                  38473         41256         +7.23%
BenchmarkProcessor/TestProcessMetricMinimalValidate-4                                 9953          10253         +3.01%
BenchmarkProcessor/TestProcessMetricMinimalDecode-4                                   4079          4492          +10.13%
BenchmarkProcessor/TestProcessMetricMinimalTransform-4                                4958          5334          +7.58%
BenchmarkProcessor/TestProcessMetricMinimalProcessRequest-4                           16782         18811         +12.09%
BenchmarkProcessor/TestProcessMetricMultipleSamplesValidate-4                         21222         24926         +17.45%
BenchmarkProcessor/TestProcessMetricMultipleSamplesDecode-4                           8218          8831          +7.46%
BenchmarkProcessor/TestProcessMetricMultipleSamplesTransform-4                        20099         25627         +27.50%
BenchmarkProcessor/TestProcessMetricMultipleSamplesProcessRequest-4                   44423         52221         +17.55%
BenchmarkBackendProcessor/TestProcessTransactionFullValidate-4                        77441         83642         +8.01%
BenchmarkBackendProcessor/TestProcessTransactionFullDecode-4                          17301         18512         +7.00%
BenchmarkBackendProcessor/TestProcessTransactionFullTransform-4                       44878         56887         +26.76%
BenchmarkBackendProcessor/TestProcessTransactionFullProcessRequest-4                  141385        159719        +12.97%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesValidate-4                  43348         44805         +3.36%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesDecode-4                    12830         13447         +4.81%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesTransform-4                 27137         32116         +18.35%
BenchmarkBackendProcessor/TestProcessTransactionNullValuesProcessRequest-4            79079         85042         +7.54%
BenchmarkBackendProcessor/TestProcessSystemNullValidate-4                             9856          10212         +3.61%
BenchmarkBackendProcessor/TestProcessSystemNullDecode-4                               4070          4297          +5.58%
BenchmarkBackendProcessor/TestProcessSystemNullTransform-4                            6566          5072          -22.75%
BenchmarkBackendProcessor/TestProcessSystemNullProcessRequest-4                       19059         19535         +2.50%
BenchmarkBackendProcessor/TestProcessProcessNullValidate-4                            10309         11035         +7.04%
BenchmarkBackendProcessor/TestProcessProcessNullDecode-4                              4117          4539          +10.25%
BenchmarkBackendProcessor/TestProcessProcessNullTransform-4                           6442          5028          -21.95%
BenchmarkBackendProcessor/TestProcessProcessNullProcessRequest-4                      18831         16807         -10.75%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanValidate-4                 12488         13216         +5.83%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanDecode-4                   5064          5279          +4.25%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanTransform-4                9131          8732          -4.37%
BenchmarkBackendProcessor/TestProcessTransactionMinimalSpanProcessRequest-4           25417         24420         -3.92%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceValidate-4              9634          10127         +5.12%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceDecode-4                3997          3945          -1.30%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceTransform-4             7232          5850          -19.11%
BenchmarkBackendProcessor/TestProcessTransactionMinimalServiceProcessRequest-4        18482         16455         -10.97%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessValidate-4              10623         11698         +10.12%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessDecode-4                4443          5062          +13.93%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessTransform-4             7729          6694          -13.39%
BenchmarkBackendProcessor/TestProcessTransactionMinimalProcessProcessRequest-4        21150         26021         +23.03%
BenchmarkBackendProcessor/TestProcessTransactionEmptyValidate-4                       10200         11827         +15.95%
BenchmarkBackendProcessor/TestProcessTransactionEmptyDecode-4                         4223          5000          +18.40%
BenchmarkBackendProcessor/TestProcessTransactionEmptyTransform-4                      7661          5773          -24.64%
BenchmarkBackendProcessor/TestProcessTransactionEmptyProcessRequest-4                 19734         21522         +9.06%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPValidate-4                 13519         11161         -17.44%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPDecode-4                   5313          4831          -9.07%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPTransform-4                7711          6682          -13.34%
BenchmarkBackendProcessor/TestProcessTransactionAugmentedIPProcessRequest-4           20287         19289         -4.92%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadValidate-4              7533          9046          +20.08%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadDecode-4                3024          3279          +8.43%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadTransform-4             6625          5829          -12.02%
BenchmarkBackendProcessor/TestProcessTransactionMinimalPayloadProcessRequest-4        15443         20521         +32.88%
BenchmarkRumProcessor/TestProcessTransactionRumValidate-4                             16243         20724         +27.59%
BenchmarkRumProcessor/TestProcessTransactionRumDecode-4                               6453          7438          +15.26%
BenchmarkRumProcessor/TestProcessTransactionRumTransform-4                            20974         21152         +0.85%
BenchmarkRumProcessor/TestProcessTransactionRumProcessRequest-4                       38553         46474         +20.55%
BenchmarkRumProcessor/TestProcessTransactionAugmentedMergeValidate-4                  8981          10567         +17.66%
BenchmarkRumProcessor/TestProcessTransactionAugmentedMergeDecode-4                    3727          3591          -3.65%
BenchmarkRumProcessor/TestProcessTransactionAugmentedMergeTransform-4                 8106          6514          -19.64%
BenchmarkRumProcessor/TestProcessTransactionAugmentedMergeProcessRequest-4            19810         17669         -10.81%
BenchmarkRumProcessor/TestProcessTransactionAugmentedValidate-4                       8867          8741          -1.42%
BenchmarkRumProcessor/TestProcessTransactionAugmentedDecode-4                         3994          3843          -3.78%
BenchmarkRumProcessor/TestProcessTransactionAugmentedTransform-4                      10124         6441          -36.38%
BenchmarkRumProcessor/TestProcessTransactionAugmentedProcessRequest-4                 22214         16197         -27.09%

benchmark                                                                             old allocs     new allocs     delta
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceValidate-4                    8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceDecode-4                      10             10             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceTransform-4                   38             28             -26.32%
BenchmarkBackendProcessor/TestProcessErrorMinimalServiceProcessRequest-4              56             46             -17.86%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessValidate-4                    14             14             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessDecode-4                      12             12             +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessTransform-4                   41             31             -24.39%
BenchmarkBackendProcessor/TestProcessErrorMinimalProcessProcessRequest-4              67             57             -14.93%
BenchmarkBackendProcessor/TestProcessErrorFullValidate-4                              80             80             +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullDecode-4                                93             93             +0.00%
BenchmarkBackendProcessor/TestProcessErrorFullTransform-4                             207            300            +44.93%
BenchmarkBackendProcessor/TestProcessErrorFullProcessRequest-4                        380            473            +24.47%
BenchmarkBackendProcessor/TestProcessErrorNullValuesValidate-4                        12             12             +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesDecode-4                          22             22             +0.00%
BenchmarkBackendProcessor/TestProcessErrorNullValuesTransform-4                       87             123            +41.38%
BenchmarkBackendProcessor/TestProcessErrorNullValuesProcessRequest-4                  121            157            +29.75%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPValidate-4                       8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPDecode-4                         15             15             +0.00%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPTransform-4                      42             32             -23.81%
BenchmarkBackendProcessor/TestProcessErrorAugmentedIPProcessRequest-4                 65             55             -15.38%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionValidate-4           0              0              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionDecode-4             9              9              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionTransform-4          36             26             -27.78%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadExceptionProcessRequest-4     45             35             -22.22%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogValidate-4                 8              8              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogDecode-4                   9              9              +0.00%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogTransform-4                36             26             -27.78%
BenchmarkBackendProcessor/TestProcessErrorMinimalPayloadLogProcessRequest-4           53             43             -18.87%
BenchmarkRumProcessor/TestProcessErrorRumValidate-4                                   72             72             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumDecode-4                                     57             57             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumTransform-4                                  752            742            -1.33%
BenchmarkRumProcessor/TestProcessErrorRumProcessRequest-4                             881            871            -1.14%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapValidate-4                             61             61             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapDecode-4                               44             44             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapTransform-4                            714            704            -1.40%
BenchmarkRumProcessor/TestProcessErrorRumNoSmapProcessRequest-4                       819            809            -1.22%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapValidate-4                       85             85             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapDecode-4                         56             56             +0.00%
BenchmarkRumProcessor/TestProcessErrorRumMinifiedSmapTransform-4                      426            416            -2.35%
…